requirements.txt
autoflake==2.3.1
black==25.1.0
click==8.2.1
flake8==7.2.0
flake8-import-order==0.18.2
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pycodestyle==2.13.0
pyflakes==3.3.2
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
scipy==1.15.3
setuptools==80.9.0
shiboken6==6.9.1
